added support for ioe entrance website

+++ src/utils/noticeManager.js
@@ -4,10 +4,7 @@ const cheerio = require('cheerio');
 
 async function fetchSavedNotices(filePath) {
     try {
-        if (!fs.existsSync(filePath)) {
-            fs.writeFileSync(filePath, '[]');
-            return [];
-        }
+        if (!fs.existsSync(filePath)) { fs.writeFileSync(filePath, '[]') };
         const data = fs.readFileSync(filePath, 'utf-8');
         if (data.length === 0) { return [] };
         return JSON.parse(data);
@@ -53,11 +50,21 @@ async function fetchCurrentNoticesIOM() {
         return currentNotices;
     }
     catch (err) {
-        console.error(err.messsage);
+        console.error(`Error fetching notices for IOM: ${err.response.statusText}`);
+        return [];
     }
 }
 
-async function fetchCurrentNoticesIOE() {
+async function fetchCurrentNoticesIOE(noticeType) {
+    if (noticeType === 'exam') {
+        return await examIOE();
+    } else if (noticeType === 'entrance') {
+        return await entranceIOE();
+    }
+}
+
+// IOE: http://exam.ioe.edu.np/
+async function examIOE() {
     try {
         const currentNotices = [];
         const result = await axios.get('http://exam.ioe.edu.np/');
@@ -73,7 +80,31 @@ async function fetchCurrentNoticesIOE() {
         return currentNotices;
     }
     catch (err) {
-        throw (err);
+        console.error(err);
+    }
+}
+
+// IOE: https://entrance.ioe.edu.np/
+async function entranceIOE() {
+    try {
+        const currentNotices = [];
+        const result = await axios.get('https://entrance.ioe.edu.np/Notice');
+        const $ = cheerio.load(result.data);
+        const table = $('.table.table-bordered tbody tr');
+        for (let i = 0; i < table.length; i++) {
+            const row = table.eq(i);
+            const fullNoticeUrl = 'https://entrance.ioe.edu.np' + row.find('td').eq(3).find('a').attr('href');
+            const fullNoticePage = await axios.get(fullNoticeUrl);
+            const $notice = cheerio.load(fullNoticePage.data);
+            const date = $notice('.label.label-info').text().trim().replace("Published Date: ", "");
+            const description = $notice('.well p').text().trim().replace(/ - Click Here.*$/, '');
+            const url = $notice('.well p a').attr('href');
+            currentNotices.push({ Date: date, Description: description, Url: url });
+        }
+        return currentNotices;
+    }
+    catch (err) {
+        console.error(err);
     }
 }
